Improve differential threshold detector (#441)
## Summary of changes

Amended the cli arguments for the detector:
- `begin-threshold`: the value that the trace differential must exceed
to begin a detection.
- `begin-duration`: the time the differential must exceed the threshold
to begin a detection.
- `end-threshold`: the value that the trace differential must fall below
to conclude a detection.
- `end-duration`: the time the differential must spend below the
threshold to conclude a detection.
- `cool-off`: the minimum time between the end and beginning of
successive detections.
- `peak-height-mode`: determines how peak heights are calculated (either
max trace value, or value at end of detection).
- `peak-height-basis`: whether peak height is recorded relative to the
trace baseline, or the pulse's baseline (pulse baseline is the trace
value at the just before the start of a detection).

### Issues.
Note there is some naming inconsistency in the use of `cool-down` for
`cool-off`. This is present in the fixed threshold detector too, so will
be addressed in a future PR.

## Instruction for review/testing

General code review.
Has been tested on simulated data.